fix(hooks): handle shell boundaries and quoted flags in bypass guard
